U_KE_ADV calculation added

+++ Codes/Budget/SM/ADV_KE_sm_plot.jl
@@ -45,8 +45,8 @@ for xn in cfg["xn_start"]:cfg["xn_end"]
 
         # Read time-averaged KE advection flux
         u_ke_mean = Float64.(open(joinpath(base2, "U_KE", "u_ke_mean_$suffix.bin"), "r") do io
-            nbytes = nx * ny * sizeof(Float64)
-            reshape(reinterpret(Float64, read(io, nbytes)), nx, ny)
+            nbytes = nx * ny * sizeof(Float32)
+            reshape(reinterpret(Float32, read(io, nbytes)), nx, ny)
         end)
